requirements.txt
autopep8==1.5.4
Django==2.2.9
django-cors-headers==3.5.0
django-filter==2.2.0
djangorestframework==3.11.0
flake8==3.8.3
importlib-metadata==2.0.0
Markdown==3.1.1
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2019.3
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.2.0
